Separated getLocalSize() into getOwnedSize() (number of primary-owned entries, not supported by Infinispan) and getLocallyStoredSize() (including backups, not supported by Coherence), and adapted the stages working with this.

+++ plugins/infinispan70/src/main/java/org/radargun/service/Infinispan70CacheInfo.java
@@ -33,8 +33,8 @@ public class Infinispan70CacheInfo extends Infinispan53CacheInfo {
       }
 
       @Override
-      public int getTotalSize() {
-         int totalSize = 0;
+      public long getTotalSize() {
+         long totalSize = 0;
          EntryIterable entryIterator = null;
          try {
             entryIterator = cache.filterEntries(AllFilter.INSTANCE);
